K8s Processor: move ip extractors into a separate file
